refactor(retrieve-county-data): move function to own file
