+++ covid-mapper/utils/index.js
@@ -1,3 +1,4 @@
 export { default as capitalize } from "./capitalize";
 export { default as cartesianCoordinateConverter } from "./cartesianCoordinateConverter";
 export { default as centroidRegion } from "./centroidRegion";
+export { default as retrieveCountyData } from "./retrieveCountyData";
